refactor: full repo reorganization
