+++ src/pymmcore_widgets/config_presets/__init__.py
@@ -1,0 +1,11 @@
+"""Widgets related to configuration groups and presets."""
+
+from ._group_preset_widget._group_preset_table_widget import GroupPresetTableWidget
+from ._objectives_pixel_configuration_widget import ObjectivesPixelConfigurationWidget
+from ._pixel_configuration_widget import PixelConfigurationWidget
+
+__all__ = [
+    "GroupPresetTableWidget",
+    "ObjectivesPixelConfigurationWidget",
+    "PixelConfigurationWidget",
+]
