+++ src/pymmcore_widgets/config_presets/_group_preset_widget/_group_preset_table_widget.py
@@ -20,10 +20,9 @@ from qtpy.QtWidgets import (
     QWidget,
 )
 
-from pymmcore_widgets._core import load_system_config
-from pymmcore_widgets._presets_widget import PresetsWidget
-from pymmcore_widgets._property_widget import PropertyWidget
-from pymmcore_widgets._util import block_core
+from pymmcore_widgets._util import block_core, load_system_config
+from pymmcore_widgets.control._presets_widget import PresetsWidget
+from pymmcore_widgets.device_properties._property_widget import PropertyWidget
 
 from ._add_group_widget import AddGroupWidget
 from ._add_preset_widget import AddPresetWidget
