+++ src/pymmcore_widgets/control/__init__.py
@@ -1,0 +1,28 @@
+"""Widgets that control various devices at runtime."""
+
+from ._camera_roi_widget import CameraRoiWidget
+from ._channel_group_widget import ChannelGroupWidget
+from ._channel_widget import ChannelWidget
+from ._exposure_widget import DefaultCameraExposureWidget, ExposureWidget
+from ._live_button_widget import LiveButton
+from ._load_system_cfg_widget import ConfigurationWidget
+from ._objective_widget import ObjectivesWidget
+from ._presets_widget import PresetsWidget
+from ._shutter_widget import ShuttersWidget
+from ._snap_button_widget import SnapButton
+from ._stage_widget import StageWidget
+
+__all__ = [
+    "CameraRoiWidget",
+    "ChannelGroupWidget",
+    "ChannelWidget",
+    "ConfigurationWidget",
+    "DefaultCameraExposureWidget",
+    "ExposureWidget",
+    "LiveButton",
+    "ObjectivesWidget",
+    "PresetsWidget",
+    "ShuttersWidget",
+    "SnapButton",
+    "StageWidget",
+]
