+++ src/pymmcore_widgets/device_properties/__init__.py
@@ -1,0 +1,7 @@
+"""Widgets related to device properties."""
+
+from ._properties_widget import PropertiesWidget
+from ._property_browser import PropertyBrowser
+from ._property_widget import PropertyWidget
+
+__all__ = ["PropertiesWidget", "PropertyBrowser", "PropertyWidget"]
