+++ src/pymmcore_widgets/device_properties/_properties_widget.py
