+++ src/pymmcore_widgets/device_properties/_property_widget.py
