+++ src/pymmcore_widgets/experimental.py
@@ -1,3 +1,5 @@
-from ._stack_viewer import StackViewer
+"""Experimental widgets."""
+
+from .views._stack_viewer import StackViewer
 
 __all__ = ["StackViewer"]
