+++ src/pymmcore_widgets/views/__init__.py
@@ -1,0 +1,5 @@
+"""View-related Widgets."""
+
+from ._image_widget import ImagePreview
+
+__all__ = ["ImagePreview"]
