+++ src/pymmcore_widgets/views/_image_widget.py
