+++ src/pymmcore_widgets/views/_stack_viewer/_datastore.py
