+++ src/pymmcore_widgets/views/_stack_viewer/_stack_viewer.py
